requirements.txt
# =============================================
# MindScribe - WORKING Requirements
# No dependency conflicts - Tested and verified
# =============================================

# CORE DJANGO & DEPENDENCIES
Django==3.2.21
asgiref==3.5.2
sqlparse>=0.2.2  # Let pip choose compatible version
pytz==2023.3

# MONGODB DATABASE
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2

# ENVIRONMENT & CONFIGURATION
python-decouple==3.8
python-dotenv==1.0.0

# SECURITY & AUTHENTICATION
bcrypt==4.0.1

# IMAGE PROCESSING
Pillow==10.0.1

# PDF GENERATION
reportlab==4.0.4
weasyprint==59.0
django-weasyprint==1.1.0
cairocffi==1.6.1
cairosvg==2.7.0
tinycss2==1.2.1
cssselect2==0.7.0
html5lib==1.1
pydyf==0.5.0
pyphen==0.14.0

# DEVELOPMENT & TESTING
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.2.7

# CODE QUALITY & FORMATTING
black==23.3.0
flake8==6.0.0
isort==5.12.0

# PRODUCTION DEPLOYMENT
gunicorn==20.1.0
whitenoise==6.4.0

# HTTP REQUESTS & APIs
requests==2.31.0

# AI & ML LIBRARIES
transformers==4.35.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
keybert==0.7.0
sentence-transformers==2.2.2
spacy==3.6.1
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.6.0/fr_core_news_md-3.6.0-py3-none-any.whl
pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
djangorestframework==3.14.0